add support for not before endpoint and refactoring the admin script
